More PT 81+ data
Should be complete now from a WHM standpoint, aside from some ability
potencies which are presumably the same as all others on the set.

Note that enemy autoattack values are slightly rounded for display
convenience, but only slightly - the enemies seem tuned to deal a base
damage of as close to some multiple of 1000 as the internal parameters
will allow.  (For the record, the precise base damage values vs. WHM
are 19983, 21016, and 22050+/-2 - I haven't confirmed the exact value
in the latter case, but it's probably 22049 following the arithmetic
sequence of the first two.)