PT 81-100 data (#118)
* PT 81-100 data

* PT mimics' pox ability is Malice, not Infatuation

Infatuation was a translation error, so this is an improvement.

* More PT 81+ data

* Confirmed PT 81+ Gnoll's enrage ability name

* Add notes on PT 81+ interruptible abilities

I'm not explicitly calling out well-known spells like Blizzard or Dark as
interruptible since spells can always be interrupted (as far as I'm aware).

* PT 91+ Acheron's Authority Slash is a full 360-degree pointblank

Or at least, it extends behind the enemy (could potentially be a side 180).

* More PT 91+ data

* Floor 90 damage zone is 90k/tick, not 100k

* Consistently refer to 2x2 sets of floor tiles as "blocks"

Better wording suggestions are also appreciated!  I imagine this is the
sort of thing where an image would help too.

* Add floor 60 One/Two-stone March safe spot note

* More PT 81+ data

Should be complete now from a WHM standpoint, aside from some ability
potencies which are presumably the same as all others on the set.

Note that enemy autoattack values are slightly rounded for display
convenience, but only slightly - the enemies seem tuned to deal a base
damage of as close to some multiple of 1000 as the internal parameters
will allow.  (For the record, the precise base damage values vs. WHM
are 19983, 21016, and 22050+/-2 - I haven't confirmed the exact value
in the latter case, but it's probably 22049 following the arithmetic
sequence of the first two.)

* More PT 91+ data

* More PT 91+ data

* Typo/wording fixes